requirements.txt
flask
flask-cors
python-dotenv
openai
requests
google-generativeai